PR #1 - code review changes #10 Refactoring
- More readable loop cursor name
- HttpMethodMapping - Use mapping rather than rely on enum ordinal
- Removed PathDefinitionImpl and RouteDefinitionImpl and respected interfaces.
Instead using concrete classes.
