+++ src/main/java/com/appsflyer/donkey/server/handler/InternalServerErrorHandler.java
@@ -5,13 +5,13 @@ import io.vertx.ext.web.RoutingContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class InternalServerErrorHandler implements Handler<RoutingContext>
-{
-  private static final Logger logger = LoggerFactory.getLogger(InternalServerErrorHandler.class.getName());
+public class InternalServerErrorHandler implements Handler<RoutingContext> {
+  
+  private static final Logger logger = LoggerFactory.getLogger(
+      InternalServerErrorHandler.class.getName());
   
   @Override
-  public void handle(RoutingContext ctx)
-  {
+  public void handle(RoutingContext ctx) {
     logger.error("Unhandled exception:", ctx.failure());
     ctx.response().setStatusCode(500);
   }
