+++ src/main/java/com/appsflyer/donkey/server/ring/handler/RingRequestAdapter.java
@@ -17,22 +17,22 @@ public class RingRequestAdapter implements RingHandler {
     RingRequestField[] fields = RingRequestField.values();
     var values = new Object[fields.length * 2];
     
-    var j = 0;
+    var valueIndex = 0;
     for (int i = 0; i < fields.length; i++) {
       var field = fields[i];
       Object v = field.from(ctx);
       if (v != null) {
-        values[j] = field.keyword();
-        values[j + 1] = v;
-        j += 2;
+        values[valueIndex] = field.keyword();
+        values[valueIndex + 1] = v;
+        valueIndex += 2;
       }
     }
     
-    if (j == values.length) {
+    if (valueIndex == values.length) {
       ctx.put(RING_HANDLER_RESULT, toPersistentMap(values));
     } else {
-      var copy = new Object[j];
-      System.arraycopy(values, 0, copy, 0, j);
+      var copy = new Object[valueIndex];
+      System.arraycopy(values, 0, copy, 0, valueIndex);
       ctx.put(RING_HANDLER_RESULT, toPersistentMap(copy));
     }
