--- src/main/java/com/appsflyer/donkey/server/route/PathDefinitionImpl.java
@@ -1,43 +1,0 @@
-/*
- * Copyright 2020 AppsFlyer
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.appsflyer.donkey.server.route;
-
-public class PathDefinitionImpl implements PathDefinition {
-  
-  private final String value;
-  private final MatchType matchType;
-  
-  PathDefinitionImpl(String value, MatchType matchType) {
-    this.value = value;
-    this.matchType = matchType;
-  }
-  
-  PathDefinitionImpl(String value) {
-    this(value, MatchType.SIMPLE);
-  }
-  
-  @Override
-  public String value() {
-    return value;
-  }
-  
-  @Override
-  public MatchType matchType() {
-    return matchType;
-  }
-  
-}
