+++ src/test/java/com/appsflyer/donkey/server/route/RouteSupplierImpl.java
@@ -32,19 +32,19 @@ public class RouteSupplierImpl implements RouteSupplier {
   @Override
   public RouteDefinition echo(Checkpoint requestsServed) {
     return RouteDefinition.create()
-                          .path("/echo")
-                          .handler(returnRequest(requestsServed));
+                           .path("/echo")
+                           .handler(returnRequest(requestsServed));
   }
   
   @Override
   public RouteDefinition postFormOrFile(Checkpoint requestsServed) {
     return RouteDefinition.create()
-                          .path("/post/form")
-                          .addMethod(POST)
-                          .addConsumes(APPLICATION_X_WWW_FORM_URLENCODED.toString())
-                          .addConsumes(MULTIPART_FORM_DATA.toString())
-                          .addConsumes(APPLICATION_OCTET_STREAM.toString())
-                          .handler(returnRequest(requestsServed));
+                           .path("/post/form")
+                           .addMethod(POST)
+                           .addConsumes(APPLICATION_X_WWW_FORM_URLENCODED.toString())
+                           .addConsumes(MULTIPART_FORM_DATA.toString())
+                           .addConsumes(APPLICATION_OCTET_STREAM.toString())
+                           .handler(returnRequest(requestsServed));
   }
   
   @Override
